requirements.txt
numpy==1.18.5
tensorflow==1.15
scikit-learn==0.22.2
scikit-image== 0.16.2
imageio==2.4.1
Pillow==7.0.0
scipy==1.4.1
tqdm== 4.41.1
albumentations==0.5.1
higher=0.2.1
functools=0.5
numpy==1.19.4
pandas==1.1.5
csv==1.0
